a22: rro_overlays: Increase 5 GHz network signal tolerance
On devices with cellular data available, I've been experiencing Wi-Fi dropouts on 5 GHz networks where it disconnects and falls back to cellular data around a RSSI of -77 dBm. While the Wi-Fi quality may not be ideal at this signal level, it is still better to stay on it than switch to cellular data because switching networks can be very disruptive to the user.

To make matters worse, the signal tends to oscillate around -77 dBm in my case, which causes it to oscillate between Wi-Fi and cellular data every few seconds. This causes far more disruptions than staying on weak Wi-Fi would.

These signal levels were measured empirically on a Pixel 5, but they should apply to most devices. 2.4 GHz values were found to be more or less accurate, but 5 GHz networks continued to work past the AOSP thresholds. The iPhone 6s was also content with these signal levels and still displayed 2 of 3 signal levels at -77 dBm.